Use OCTAVE_LOCAL_BUFFER for temp mem allocation
* DESCRIPTION: bump version 1.4.0+

* NEWS: note change

* src/poll.cc, src/recv.cc, src/send.cc: use OCTAVE_LOCAL_BUFFER

* Makefile: dont error on clean when no target dir

+++ Makefile
@@ -266,7 +266,7 @@ clean: clean-tarballs clean-unpacked-release clean-install clean-docs
 	test -e inst/test && rmdir inst/test || true
 	test -e $(target_dir)/fntests.log && rm -f $(target_dir)/fntests.log || true
 	@echo "## Removing target directory (if empty)..."
-	-rmdir $(target_dir)
+	test -e $(target_dir) && rmdir $(target_dir) || true
 	@echo
 	@echo "## Cleaning done"
 	@echo
